// frontend/src/backend/database.types.ts
export type Database = {
// ...
  public: {
// ...
    Functions: {
// ...
      get_student_by_id: {
        Args: {
          p_id: number;
        };
        Returns: {
          id: number;
          generation: number;
          khmer_name: string;
          english_name: string;
          date_of_birth: string;
          sex: string;
          nationality: string;
          address: string;
          phone_number: string;
          facebook_or_email: string;
          father_name: string;
          father_nationality: string;
          father_occupation: string;
          father_phone_number: string;
          mother_name: string;
          mother_nationality: string;
          mother_occupation: string;
          mother_phone_number: string;
          session: string;
          session_id: number;
          subject: string;
          subject_id: number;
          college: string;
          college_id: number;
          grade: string;
          grade_id: number;
          family_situation: string;
          family_situation_id: number;
          student_job: string;
          student_job_id: number;
        }[];
      };
// ...
    };
    Enums: {
      [_ in never]: never;
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
};
type PublicSchema = Database[Extract<keyof Database, "public">];
// ...
export type Functions<
  PublicFunctionNameOrOptions extends
    | keyof PublicSchema["Functions"]
    | { schema: keyof Database },
  FunctionName extends PublicFunctionNameOrOptions extends {
    schema: keyof Database;
  }
    ? keyof Database[PublicFunctionNameOrOptions["schema"]]["Functions"]
    : never = never,
> = PublicFunctionNameOrOptions extends { schema: keyof Database }
  ? Database[PublicFunctionNameOrOptions["schema"]]["Functions"][FunctionName] extends {
      Returns: infer R;
    }
    ? R
    : never
  : PublicFunctionNameOrOptions extends keyof PublicSchema["Functions"]
    ? PublicSchema["Functions"][PublicFunctionNameOrOptions] extends {
        Returns: infer R;
      }
      ? R
      : never
    : never;